separated out planner, renamed chat methods
